+++ packages/orq-rc/src/models/operations/createdatasetitem.ts
@@ -2607,7 +2607,7 @@ export const Evaluations3$inboundSchema: z.ZodType<
   explanation: z.string().optional(),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2026-07-05T09:16:37.079Z",
+    "2026-07-06T04:42:26.532Z",
   ).transform(v => new Date(v)),
   type: z.literal("string_array"),
   values: z.array(z.string()),
@@ -2692,7 +2692,7 @@ export const Evaluations2$inboundSchema: z.ZodType<
   explanation: z.string().optional(),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2026-07-05T09:16:37.078Z",
+    "2026-07-06T04:42:26.531Z",
   ).transform(v => new Date(v)),
   type: z.literal("number"),
   value: z.number(),
@@ -2777,7 +2777,7 @@ export const Evaluations1$inboundSchema: z.ZodType<
   explanation: z.string().optional(),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2026-07-05T09:16:37.077Z",
+    "2026-07-06T04:42:26.530Z",
   ).transform(v => new Date(v)),
   type: z.literal("string"),
   value: z.string(),
@@ -2862,7 +2862,7 @@ export const CreateDatasetItemResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-05T09:16:20.541Z",
+    "2026-07-06T04:42:03.033Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {