+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -412,7 +412,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KWRS1F8AN0NDGB5WRGXFJKB8"),
+  _id: z.string().default("01KWTVR58GMARGV79KBEAK8MG5"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: CreateDatasourceStatus$inboundSchema,