+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-05T09:16:20.541Z",
+    "2026-07-06T04:42:03.033Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {