+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -182,7 +182,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KWRS1F7RHJJDD3ZZGJGPEX1F"),
+  _id: z.string().default("01KWTVR58DV1G7VEBE32FFF0WG"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: ListDatasourcesStatus$inboundSchema,