+++ packages/orq-rc/RELEASES.md
@@ -10579,3 +10579,13 @@ Based on:
 - [typescript v4.12.0-rc.1] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.1 - packages/orq-rc
+
+## 2026-07-06 04:44:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.788.1 (2.915.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.2] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.2 - packages/orq-rc